0.3.5 Bot - Refactor #3 Final
